Cleanup a lot of warnings

+++ ConsoleAppFW_test/consoleapp_test.cpp
@@ -39,22 +39,22 @@ using MyAppTestDeath = MyAppTest;
 TEST_F(MyAppTest, Wrong_Syntax)
 {
 	std::vector<char*> argv{ "program.exe", "/t" };
-	EXPECT_STREQ(cons0.Arguments(argv.size(), &argv[0]).c_str(), "Unknown argument '/t' - see program.exe /? for help.");
-	EXPECT_STREQ(cons1.Arguments(argv.size(), &argv[0]).c_str(), "Unknown argument '/t' - see program.exe /? for help.");
+	EXPECT_STREQ(cons0.Arguments((int)argv.size(), &argv[0]).c_str(), "Unknown argument '/t' - see program.exe /? for help.");
+	EXPECT_STREQ(cons1.Arguments((int)argv.size(), &argv[0]).c_str(), "Unknown argument '/t' - see program.exe /? for help.");
 }
 
 TEST_F(MyAppTestDeath, Parse_More_Than_One_Time)
 {
 	std::vector<char*> argv{ "program.exe", "/t" };
-	cons0.Arguments(argv.size(), &argv[0]);
-	EXPECT_DEATH(cons0.Arguments(argv.size(), &argv[0]), "");
+	cons0.Arguments((int)argv.size(), &argv[0]);
+	EXPECT_DEATH(cons0.Arguments((int)argv.size(), &argv[0]), "");
 }
 
 TEST_F(MyAppTest, Help)
 {
 	std::vector<char*> argv{ "program.exe", "/?" };
-	EXPECT_STREQ(cons0.Arguments(argv.size(), &argv[0]).c_str(), "?");
-	EXPECT_STREQ(cons1.Arguments(argv.size(), &argv[0]).c_str(), "?");
+	EXPECT_STREQ(cons0.Arguments((int)argv.size(), &argv[0]).c_str(), "?");
+	EXPECT_STREQ(cons1.Arguments((int)argv.size(), &argv[0]).c_str(), "?");
 }
 
 TEST_F(MyAppTestDeath, Run_Without_Parsing_Arguments)
@@ -65,14 +65,14 @@ TEST_F(MyAppTestDeath, Run_Without_Parsing_Arguments)
 TEST_F(MyAppTest, No_Matching_File)
 {
 	std::vector<char*> argv{ "program.exe", "H:/Windows/System32/msxml2.*" };
-	EXPECT_STREQ(cons0.Arguments(argv.size(), &argv[0]).c_str(), "Arguments are checked.");
+	EXPECT_STREQ(cons0.Arguments((int)argv.size(), &argv[0]).c_str(), "Arguments are checked.");
 	EXPECT_ANY_THROW(cons0.Run());
 }
 
 TEST_F(MyAppTest, Matching_Files)
 {
 	std::vector<char*> argv{ "program.exe", "H:/Windows/System32/msxml?.*" };
-	EXPECT_STREQ(cons0.Arguments(argv.size(), &argv[0]).c_str(), "Arguments are checked.");
+	EXPECT_STREQ(cons0.Arguments((int)argv.size(), &argv[0]).c_str(), "Arguments are checked.");
 	auto files = cons0.values("file");
 	EXPECT_EQ(files.size(), 1);
 }
@@ -85,8 +85,8 @@ TEST_F(MyAppTestDeath, Get_Values_Before_Parsing_Arguments)
 TEST_F(MyAppTest, Run_Test)
 {
 	std::vector<char*> argv{ "program.exe", "H:/Windows/System32/msxml?.*" };
-	EXPECT_STREQ(cons0.Arguments(argv.size(), &argv[0]).c_str(), "Arguments are checked.");
-	EXPECT_STREQ(cons1.Arguments(argv.size(), &argv[0]).c_str(), "Arguments are checked.");
+	EXPECT_STREQ(cons0.Arguments((int)argv.size(), &argv[0]).c_str(), "Arguments are checked.");
+	EXPECT_STREQ(cons1.Arguments((int)argv.size(), &argv[0]).c_str(), "Arguments are checked.");
 	EXPECT_EQ(cons0.Run(), 2);
 	EXPECT_EQ(cons1.Run(), 2);
 }
